5-aml-pytorch-deployment-samples/app-service-code/requirements.txt
fastapi
gunicorn
uvicorn
python-multipart==0.0.5
torch==1.9.0
pillow==8.3.2
azureml-defaults==1.35.0
